AIO update v0.2
everything works MUCH better now, SSA is in swal now, the swal is in the same style as the rest of the site, the functionality is still the same and everything works.
TODO: make the login page as well and make sure it works smoothly.
maybe put limits on inputs, so username can't be 9999999999 characters long.